requirements.txt
scipy
scikits.talkbox
pybrain
pybrain2
numpy
pydub
